created "modals" folder and moved some components
