+++ src/components/modals/SignIn/SignIn.jsx
@@ -4,20 +4,16 @@ import * as Yup from 'yup';
 import { FaSpinner } from 'react-icons/fa';
 
 import Modal from '../Modal/Modal';
-import FormField from '../FormField/FormField';
+import FormField from '../../FormField/FormField';
 
 const SignInSchema = Yup.object().shape({
   username: Yup.string().min(3).max(24).required(),
-  email: Yup.string().email().required(),
   password: Yup.string().min(8).max(32).required(),
-  passwordConfimation: Yup.string()
-    .oneOf([Yup.ref('password'), null], 'Passwords does not match')
-    .required(),
 });
 
-function SignUp({ open, setOpen, setSignIn }) {
+function SignIn({ open, setOpen, setSignUp }) {
   return (
-    <Modal label="Sign up" open={open} setOpen={setOpen}>
+    <Modal label="Sign in" open={open} setOpen={setOpen}>
       <Formik
         initialValues={{
           username: '',
@@ -48,29 +44,11 @@ function SignUp({ open, setOpen, setSignIn }) {
               type="text"
             />
             <FormField
-              autoComplete="email"
-              className="mb-3"
-              errors={errors}
-              label="Email address"
-              name="email"
-              touched={touched}
-              type="email"
-            />
-            <FormField
-              autoComplete="new-password"
+              autoComplete="password"
               className="mb-3"
               errors={errors}
               label="Password"
               name="password"
-              touched={touched}
-              type="password"
-            />
-            <FormField
-              autoComplete="new-password"
-              className="mb-3"
-              errors={errors}
-              label="Repeat password"
-              name="passwordConfimation"
               touched={touched}
               type="password"
             />
@@ -82,28 +60,28 @@ function SignUp({ open, setOpen, setSignIn }) {
                   disabled
                 >
                   <FaSpinner className="animate-spin -ml-1 mr-3 h-5 w-5 text-white" />
-                  Signing up...
+                  Signing in...
                 </button>
               ) : (
                 <button
                   type="submit"
                   className="text-lg text-gray-100 bg-red-600 hover:bg-red-700 transition rounded px-5 py-2"
                 >
-                  Sign up
+                  Sign in
                 </button>
               )}
               <p className="text-right text-xs text-gray-500 mb-0">
-                Already have an account?
+                Don&apos;t have an account yet?
                 <br />
                 <button
                   type="button"
                   className="text-sm text-red-600 hover:text-red-700 transition"
                   onClick={() => {
                     setOpen(false);
-                    setSignIn(true);
+                    setSignUp(true);
                   }}
                 >
-                  Sign in
+                  Sign up
                 </button>
               </p>
             </div>
@@ -113,10 +91,10 @@ function SignUp({ open, setOpen, setSignIn }) {
     </Modal>
   );
 }
-SignUp.propTypes = {
+SignIn.propTypes = {
   open: PropTypes.bool.isRequired,
   setOpen: PropTypes.func.isRequired,
-  setSignIn: PropTypes.func.isRequired,
+  setSignUp: PropTypes.func.isRequired,
 };
 
-export default SignUp;
+export default SignIn;
